C# InitParams is now a normal managed object with property accessors.

+++ Upnp/Public/Cs/TestProxy.cs
@@ -13,7 +13,7 @@ namespace Zapp
             Core.Library lib = new Core.Library();
             lib.Initialise(initParams);
             lib.StartCp();
-            new Runner(initParams.MsearchTimeSecs());
+            new Runner((int)initParams.MsearchTimeSecs);
             lib.Close();
         }
     }
